+++ examples/workflow/nuxt/server/api/sleep.ts
@@ -1,5 +1,5 @@
 
-import { serve } from "@upstash/qstash/workflow/nuxt";
+import { serve } from "@upstash/qstash/nuxt";
 
 const someWork = (input: string) => {
   return `processed '${JSON.stringify(input)}'`
