+++ examples/workflow/sveltekit/src/routes/sleepWithoutAwait/+server.ts
@@ -1,4 +1,4 @@
-import { serve } from "@upstash/qstash/workflow/sveltekit";
+import { serve } from "@upstash/qstash/svelte";
 import { env } from '$env/dynamic/private'
 import { Client } from "@upstash/qstash";
