+++ platforms/nextjs.ts
@@ -4,6 +4,9 @@ import { type NextRequest } from "next/server";
 import { type NextFetchEvent, NextResponse } from "next/server";
 import { Receiver } from "../src/receiver";
 
+import type { WorkflowServeParameters } from "../src/client/workflow";
+import { serve as serveBase } from "../src/client/workflow";
+
 export type VerifySignatureConfig = {
   currentSigningKey?: string;
   nextSigningKey?: string;
@@ -191,3 +194,19 @@ export function verifySignatureAppRouter(
     return handler(request as NextRequest);
   };
 }
+
+export const serve = <TInitialPayload = unknown>({
+  routeFunction,
+  options,
+}: WorkflowServeParameters<TInitialPayload, NextResponse>): ((
+  request: NextRequest
+) => Promise<NextResponse>) => {
+  return serveBase<TInitialPayload, NextRequest, NextResponse>({
+    routeFunction,
+    options: {
+      onStepFinish: (workflowRunId: string) =>
+        new NextResponse(JSON.stringify({ workflowRunId }), { status: 200 }),
+      ...options,
+    },
+  });
+};
