+++ platforms/solidjs.ts
@@ -1,6 +1,9 @@
 import type { APIEvent, APIHandler } from "@solidjs/start/server";
 import { Receiver } from "../src";
 
+import type { WorkflowServeParameters } from "../src/client/workflow";
+import { serve as serveBase } from "../src/client/workflow";
+
 type VerifySignatureConfig = {
   currentSigningKey?: string;
   nextSigningKey?: string;
@@ -43,4 +46,29 @@ export const verifySignatureSolidjs = (
     }
     return handler(event);
   };
+};
+
+export const serve = <TInitialPayload = unknown>({
+  routeFunction,
+  options,
+}: WorkflowServeParameters<TInitialPayload>) => {
+  // Create a handler which receives an event and calls the
+  // serveBase method
+  const handler = async (event: APIEvent) => {
+    // verify that the request is POST
+    const method = event.request.method;
+    if (method.toUpperCase() !== "POST") {
+      return new Response("Only POST requests are allowed in worklfows", { status: 405 });
+    }
+
+    // create serve handler
+    const serveHandler = serveBase<TInitialPayload>({
+      routeFunction,
+      options,
+    });
+
+    // invoke serve handler and return result
+    return serveHandler(event.request);
+  };
+  return handler;
 };
