--- src/client/workflow/frameworks/nextjs.ts
@@ -1,20 +1,0 @@
-import type { NextRequest } from "next/server";
-import { NextResponse } from "next/server";
-import type { WorkflowServeParameters } from "../types";
-import { serve as serveBase } from "../serve";
-
-export const serve = <TInitialPayload = unknown>({
-  routeFunction,
-  options,
-}: WorkflowServeParameters<TInitialPayload, NextResponse>): ((
-  request: NextRequest
-) => Promise<NextResponse>) => {
-  return serveBase<TInitialPayload, NextRequest, NextResponse>({
-    routeFunction,
-    options: {
-      onStepFinish: (workflowRunId: string) =>
-        new NextResponse(JSON.stringify({ workflowRunId }), { status: 200 }),
-      ...options,
-    },
-  });
-};
